feat: add admin router with user and file management endpoints
