+++ server/services/__init__.py
@@ -1,2 +1,3 @@
 from .auth.router import router as auth_router
 from .file.router import router as file_router
+from .admin.router import router as admin_router
