+++ server/services/admin/__init__.py
@@ -1,0 +1,1 @@
+from .router import router as admin_router
