+++ server/services/auth/__init__.py
@@ -1,0 +1,1 @@
+from .router import router as auth_router
